Trivia Game
I created this simple trivia game to practice basic C++ concepts like input handling and conditional statements. In this game, users are presented with three multiple-choice questions(the code can be replicated and modified for more questions. They need to enter the number corresponding to the correct answer. After each question, the program provides feedback on whether the answer was correct or incorrect. At the end, the user's score is displayed.